Fix wrench build on Windows
